fixed component and updated other files accordingly
